requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-CORS==4.0.0
Flask-Session==0.5.0
psycopg2-binary==2.9.7
python-dotenv==1.0.0
numpy==1.24.4
pandas==2.0.3
openpyxl==3.1.2
xlrd==2.0.1
alembic==1.16.4
